requirements.txt
watchdog
google-genai
reportlab
pillow